pyproject.toml
[project]
name = "dns-manager"
dynamic = ["version"]
description = "Manage your DNS records with ease."
authors = [
    {name = "Zhan Rongrui", email = "2742392377@qq.com"},
]
dependencies = [
    "tencentcloud-sdk-python-dnspod",
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "pyparsing>=3.1.1",
    "typer[all]>=0.9.0",
    "cachetools>=5.3.2",
    "pydantic>=2.5.3",
    "auto-token>=0.0.4",
    "PyYaml>=6.0.1",
    "toml>=0.10.2",
    "httpx>=0.26.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
watch = ["watchfiles"]
all = ["dns-manager[watch]"]

[project.urls]
Homepage = "https://github.com/zrr1999/dns-manager.git"
Repository = "https://github.com/zrr1999/dns-manager.git"

[project.scripts]
dns-manager = "dns_manager.__main__:app"
dnsm = "dns_manager.__main__:app"

[tool.pdm.version]
source = "scm"
write_to = "dns_manager/__version__.py"
write_template = "__version__ = '{}'"

[tool.pdm.dev-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "coverage",
]

[tool.pdm.build]
includes = ["dns_manager"]

[tool.ruff]
line-length = 100
src = ["dns_manager", "tests"]
target-version = "py311"

[tool.ruff.lint]
select = [
  "F",
  "SIM",
  "UP",
  "FA",   # flake8-annotations
  "I",    # isort
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "PGH",  # pygrep-hooks
  "RUF",  # ruff
  "E",    # pycodestyle
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]

[tool.ruff.lint.isort]
known-first-party = ["dns_manager"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "I002"]

[build-system]
requires = ["pdm-backend", "pdm-pyfuture"]
build-backend = "pdm.backend"
